requirements.txt
click==8.1.8
colorama==0.4.6
iniconfig==2.1.0
-e git+ssh://git@github.com/BenKorea/nmdose.git@cc387ea82f67a8041e232f2db24292b3fa925df5#egg=nmdose
packaging==25.0
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydicom==3.0.1
pytest==8.3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
setuptools==80.7.1
six==1.17.0
wheel==0.45.1
